Crate alac

Structs

Decoder
An ALAC packet decoder.

InvalidData
An error indicating user-provided data is invalid.

Packets
An iterator-like type that decodes packets into a user-provided buffer.

Reader
An ALAC reader and decoder supporting mp4 and caf files (if the respective Cargo features are enabled).

Samples
An iterator that yields samples of type S read from a Reader.

StreamInfo
Codec initialisation parameters for an ALAC stream.

Enums

Format
The format of an ALAC file.

ReadError
An error when reading an ALAC file using a Reader.

Traits

Sample
A type that can be used to represent audio samples.